💥 [GH-0] redesigned structure of di feature and renamed to context
