--- library/src/main/kotlin/io/mzlnk/javalin/xt/di/JavalinDIException.kt
@@ -1,13 +1,0 @@
-package io.mzlnk.javalin.xt.di
-
-/**
- * Exception thrown when an error related to Javalin DI occurs.
- */
-class JavalinDIException(message: String) : RuntimeException(message) {
-
-    /*
-     * Prevent the stack trace from being filled in.
-     */
-    override fun fillInStackTrace(): Throwable = this
-
-}
