--- library/src/main/kotlin/io/mzlnk/javalin/xt/di/Singleton.kt
@@ -1,10 +1,0 @@
-package io.mzlnk.javalin.xt.di
-
-/**
- * Marks a function as a singleton.
- *
- * Singletons are used to define the singletons that will be managed by the DI context.
- */
-@Retention(AnnotationRetention.SOURCE)
-@Target(AnnotationTarget.FUNCTION)
-annotation class Singleton
