+++ library/src/main/kotlin/io/mzlnk/javalin/xt/internal/JavalinXtProxy.kt
@@ -15,7 +15,7 @@ import io.javalin.security.RouteRole
 import io.javalin.websocket.WsConfig
 import io.javalin.websocket.WsExceptionHandler
 import io.javalin.websocket.WsHandlerType
-import io.mzlnk.javalin.xt.di.context.JavalinContext
+import io.mzlnk.javalin.xt.context.ApplicationContext
 import io.mzlnk.javalin.xt.properties.ApplicationProperties
 import jakarta.servlet.Servlet
 import java.util.function.Consumer
@@ -28,7 +28,7 @@ import java.util.function.Consumer
  */
 internal class JavalinXtProxy(
     javalin: Javalin,
-    val context: JavalinContext,
+    val context: ApplicationContext,
     val properties: ApplicationProperties
 ) : Javalin(null) {
