+++ library/src/main/kotlin/io/mzlnk/javalin/xt/internal/context/SingletonDefinitionSource.kt
@@ -1,0 +1,22 @@
+package io.mzlnk.javalin.xt.internal.context
+
+import io.mzlnk.javalin.xt.context.definition.SingletonDefinition
+import io.mzlnk.javalin.xt.context.definition.SingletonDefinitionProvider
+import java.util.*
+
+// TODO: refactor it
+
+internal fun interface SingletonDefinitionSource {
+
+    fun definitions(): List<SingletonDefinition<*>>
+
+}
+
+internal object DefaultSingletonDefinitionSource : SingletonDefinitionSource {
+
+    override fun definitions(): List<SingletonDefinition<*>> {
+        val providers = ServiceLoader.load(SingletonDefinitionProvider::class.java).toList()
+        return providers.flatMap { it.definitions }
+    }
+
+}
