+++ library/src/main/kotlin/io/mzlnk/javalin/xt/javalin-xt.kt
@@ -1,12 +1,12 @@
 package io.mzlnk.javalin.xt
 
 import io.javalin.Javalin
-import io.mzlnk.javalin.xt.di.context.DefaultJavalinContext
-import io.mzlnk.javalin.xt.di.context.JavalinContext
-import io.mzlnk.javalin.xt.di.context.NoOpJavalinContext
+import io.mzlnk.javalin.xt.internal.context.DefaultApplicationContext
+import io.mzlnk.javalin.xt.context.ApplicationContext
+import io.mzlnk.javalin.xt.internal.context.NoOpApplicationContext
 import io.mzlnk.javalin.xt.internal.JavalinXtProxy
-import io.mzlnk.javalin.xt.internal.di.context.DefaultSingletonDefinitionSource
-import io.mzlnk.javalin.xt.internal.di.context.SingletonDefinitionContext
+import io.mzlnk.javalin.xt.internal.context.DefaultSingletonDefinitionSource
+import io.mzlnk.javalin.xt.internal.context.SingletonDefinitionContext
 import io.mzlnk.javalin.xt.internal.properties.ApplicationPropertiesFactory
 import io.mzlnk.javalin.xt.properties.ApplicationProperties
 import org.slf4j.LoggerFactory
@@ -23,7 +23,7 @@ fun Javalin.xt(init: JavalinXtConfiguration.() -> Unit = {}): Javalin {
     val context = config.di
         .takeIf { it.enabled }
         ?.let { defaultDiContext() }
-        ?: NoOpJavalinContext
+        ?: NoOpApplicationContext
 
     val properties = ApplicationPropertiesFactory().create(config.properties)
 
@@ -34,7 +34,7 @@ fun Javalin.xt(init: JavalinXtConfiguration.() -> Unit = {}): Javalin {
     )
 }
 
-private fun defaultDiContext(): JavalinContext {
+private fun defaultDiContext(): ApplicationContext {
     val (context, elapsedTime) = measureTimedValue {
         val definitions = DefaultSingletonDefinitionSource.definitions()
         SingletonDefinitionContext.create(definitions)
@@ -42,7 +42,7 @@ private fun defaultDiContext(): JavalinContext {
 
     LOG.info("DI context created in $elapsedTime \\o/. Loaded ${context.size()} singletons")
 
-    return DefaultJavalinContext.create(context)
+    return DefaultApplicationContext.create(context)
 }
 
 
@@ -51,7 +51,7 @@ private fun defaultDiContext(): JavalinContext {
  *
  * @return the context
  */
-val Javalin.context: JavalinContext
+val Javalin.context: ApplicationContext
     get() {
         if (this !is JavalinXtProxy) {
             throw IllegalStateException("This is javalin-xt feature which has not been enabled. Call Javalin.xt() first.")
