+++ library/src/test/kotlin/io/mzlnk/javalin/xt/type/TypeReferenceTest.kt
@@ -1,6 +1,6 @@
 package io.mzlnk.javalin.xt.type
 
-import io.mzlnk.javalin.xt.di.type.TypeReference
+import io.mzlnk.javalin.xt.context.TypeReference
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
